requirements.txt
numpy~=1.19.5
tensorflow~=2.4.1
msgpack~=1.0.2
six~=1.15.0
Pillow~=8.0.1
requests~=2.24.0
opencv-python~=4.4.0.46
vidgear~=0.1.9
onnxruntime~=1.5.2